+++ tests/data/sphinx/no_return_section/cases.py
@@ -67,3 +67,42 @@ def func7(text: str) -> NoReturn:
     :type text: str
     """
     exit(1)
+
+
+def func8(num: int) -> Generator[None, None, None]:
+    """
+    Do something
+
+    :param num: A number
+    :type num: int
+    """
+    for i in range(num):
+        yield i
+
+    return 'All numbers yielded!'
+
+
+def func9(num: int) -> Generator[None, None, NoReturn]:
+    """
+    Do something
+
+    :param num: A number
+    :type num: int
+    """
+    for i in range(num):
+        yield i
+
+    return 'All numbers yielded!'
+
+
+def func10(num: int) -> Generator[bool, None, str]:
+    """
+    Do something
+
+    :param num: A number
+    :type num: int
+    """
+    for i in range(num):
+        yield i
+
+    return 'All numbers yielded!'